+++ index.html
@@ -249,6 +249,6 @@
   </template>
 
 </body>
-<script src="./js/main.js"></script>
+<script type="module" src="./js/main.js"></script>
 
 </html>
